+++ patzilla/util/image/convert.py
@@ -1,70 +1,80 @@
 # -*- coding: utf-8 -*-
-# (c) 2011-2016 Andreas Motl <andreas.motl@elmyra.de>
+# (c) 2011-2018 Andreas Motl <andreas.motl@ip-tools.org>
 import os
+import where
 import logging
+import datetime
 import StringIO
 import subprocess
-import datetime
+from six import BytesIO
 from tempfile import NamedTemporaryFile
 from cornice.util import to_list
 from patzilla.util.python.decorators import memoize
+from patzilla.util.python.system import run_command
 
-log = logging.getLogger(__name__)
+logger = logging.getLogger(__name__)
 
-def gif_to_tiff(payload):
 
-    # debugging
-    delete = True
-
-    infile = NamedTemporaryFile(prefix='tmp-gif_to_tiff-', delete=delete)
-    infile.write(payload)
-    infile.flush()
-
-    outfile_name = infile.name + '-out'
-
-    command = ['gif2tiff', infile.name, outfile_name]
-    command_string = ' '.join(command)
-
-    proc = subprocess.Popen(
-        command,
-        shell = (os.name == 'nt'),
-        stdin = subprocess.PIPE,
-        stdout = subprocess.PIPE,
-        stderr = subprocess.PIPE,
-    )
-
-    stdout = stderr = ''
-
-    try:
-        stdout, stderr = proc.communicate()
-        if proc.returncode is not None and proc.returncode != 0:
-            raise Exception('GIF to TIFF conversion failed')
-
-        payload = file(outfile_name, 'rb').read()
-        os.unlink(outfile_name)
-        return payload
-
-    except:
-        log.error('GIF to TIFF conversion failed. returncode={returncode}, command="{command_string}", stdout={stdout}, stderr={stderr}'.format(returncode=proc.returncode, **locals()))
-        raise Exception('GIF to TIFF conversion failed')
-
-def to_png(tiff_payload, format='tif'):
-
-    # unfortunately, PIL can not handle G4 compression ...
-    # Failure: exceptions.IOError: decoder group4 not available
-    # maybe patch: http://mail.python.org/pipermail/image-sig/2003-July/002354.html
+def to_png(tiff, width=None, height=None):
     """
-    import Image
-    png = StringIO.StringIO()
+    Convert image to PNG format with optional resizing.
+
+    :param tiff: A stream buffer object like BytesIO
+    :param width: The width of the image in pixels (optional)
+    :param height: The height of the image in pixels (optional)
+    :return: A BytesIO object instance containing image data
+    """
+
+
+    """
+    The PIL module didn't properly support TIFF images with G4 compression::
+
+        Failure: exceptions.IOError: decoder group4 not available
+        Maybe patch: http://mail.python.org/pipermail/image-sig/2003-July/002354.html
+
+    Nowadays, this should be supported by Pillow on recent platforms:
+    https://pillow.readthedocs.io/en/latest/releasenotes/5.0.0.html#compressed-tiff-images
+    """
     try:
-        Image.open(StringIO.StringIO(tiff_payload)).save(png, 'PNG')
+        from PIL import Image
+
+        # Read image
+        image = Image.open(tiff)
+
+        if width and height:
+
+            # Convert image to grayscale
+            image = image.convert('L')
+
+            # Resize image
+            image.thumbnail((width, height), Image.LANCZOS)
+
+        # Save image into a stream buffer
+        png = BytesIO()
+        image.save(png, 'PNG')
+
+        # Readers should start reading at the beginning of the stream
         png.seek(0)
-    except Exception, e:
-        print "ERROR (PIL+G4)!", e
-        pass
-    """
+
+        return png
+
+    except Exception as ex:
+        logger.warning('Image conversion using "Pillow" failed: {}'.format(ex))
+
 
     """
+    However, if the conversion using "Pillow" fails for some reason,
+    let's try to use the "convert" utility from ImageMagick.
+
+
+    Instructions for installing ImageMagick on Debian::
+
+        apt install imagemagick
+
+    Instructions for installing ImageMagick on Windows::
+
+        https://www.imagemagick.org/script/download.php#windows
+
     Instructions for building ImageMagick on Debian::
 
         # https://packages.debian.org/source/wheezy/imagemagick
@@ -77,51 +87,74 @@ def to_png(tiff_payload, format='tif'):
 
     """
 
+    more_args = []
 
-    # ... so use ImageMagick! ;-(
-    # http://www.imagemagick.org/pipermail/magick-users/2003-May/008869.html
-    #convert_bin = os.path.join(os.path.dirname(__file__), 'imagemagick', 'convert.exe')
+    # Compute value for "resize" parameter
+    size = ''
+    if width or height:
+
+        if width:
+            size += str(width)
+
+        # Use "x" for separating "width" and "height" when resizing
+        size += 'x'
+
+        if height:
+            size += str(height)
+
+        more_args += ['-resize', size]
 
     convert = find_convert()
-    #command = [convert, 'tif:-', '+set', 'date:create', '+set', 'date:modify', 'png:-']
+    if not convert:
+        message = 'Could not find ImageMagick program "convert", please install from e.g. https://imagemagick.org/'
+        logger.error(message)
+        raise AssertionError(message)
+
     command = [
-                convert,
-                #'{0}:-'.format(format),
-                '-',                            # Convert from any format
-                '+set', 'date:create', '+set', 'date:modify',
-                # FIXME: make this configurable
-                '-resize', '530x',
-                '-colorspace', 'rgb', '-flatten', '-depth', '8',
-                '-antialias', '-quality', '100', '-density', '300',
-                #'-level', '30%,100%',
-                'png:-']
+        convert,
+        '+set', 'date:create', '+set', 'date:modify',
+        '-colorspace', 'rgb', '-flatten', '-depth', '8',
+        '-antialias', '-quality', '100', '-density', '300',
+        # '-level', '30%,100%',
 
-    command_debug = ' '.join(command)
+        # Debugging
+        # (see "convert -list debug")
+        #'-verbose',
+        #'-debug', 'All',
 
-    proc = subprocess.Popen(
-        command,
-        shell = (os.name == 'nt'),
-        #shell = True,
-        stdin = subprocess.PIPE,
-        stdout = subprocess.PIPE,
-        stderr = subprocess.PIPE,
-    )
+        ] \
+        + more_args + \
+        [
 
-    stdout = stderr = ''
+        # Convert from specific format
+        #'{0}:-'.format(format),
 
+        # Convert from any format
+        '-',
+
+        # Convert to PNG format
+        'png:-',
+    ]
+
+    command_string = ' '.join(command)
     try:
-        stdout, stderr = proc.communicate(tiff_payload)
-        if proc.returncode is not None and proc.returncode != 0:
-            raise Exception('TIFF to PNG conversion failed')
-    except:
-        log.error('TIFF to PNG conversion failed, {1}. returncode={2}, command="{0}"'.format(command_debug, stderr, proc.returncode))
-        raise Exception('TIFF to PNG conversion failed')
+        logger.debug('Converting image using "{}"'.format(command_string))
+        return run_imagemagick(command, tiff.read())
 
-    if 'ImageMagick' in stdout[:200]:
-        log.error('TIFF to PNG conversion failed, stdout={1}, stderr={1}. command="{0}"'.format(command_debug, stdout, stderr))
-        raise Exception('TIFF to PNG conversion failed')
+    except Exception as ex:
+        logger.error('Image conversion using ImageMagicks "convert" program failed: {}'.format(ex))
+        raise
 
-    return stdout
+
+def run_imagemagick(command, input=None):
+    output = run_command(command, input)
+    if 'ImageMagick' in output.read()[:200]:
+        command_string = ' '.join(command)
+        message = 'Image conversion failed, found "ImageMagick" in STDOUT. Command was "{}"'.format(command_string)
+        logger.error(message)
+        raise RuntimeError(message)
+    output.seek(0)
+    return output
 
 
 def png_resize(png_payload, width):
@@ -163,9 +196,14 @@ def pdf_join(pages):
     # pdftk in.pdf update_info_utf8 in.info output out.pdf
     # pdftk in.pdf attach_files table1.html table2.html to_page 6 output out.pdf
 
-    command = [
-        find_pdftk(),
-    ]
+    pdftk = find_pdftk()
+    if not pdftk:
+        message = 'Could not find program "pdftk", please install it'
+        logger.error(message)
+        raise AssertionError(message)
+
+    # Build shellout command
+    command = [pdftk]
     tmpfiles = []
     for page in pages:
         tmpfile = NamedTemporaryFile()
@@ -177,7 +215,7 @@ def pdf_join(pages):
 
     command += ['cat', 'output', '-']
 
-    #log.info('command={0}'.format(' '.join(command)))
+    #logger.info('command={0}'.format(' '.join(command)))
 
     cmddebug = ' '.join(command)
     stdout = stderr = ''
@@ -194,10 +232,10 @@ def pdf_join(pages):
 
         stdout, stderr = proc.communicate()
         if proc.returncode is not None and proc.returncode != 0:
-            log.error('pdftk joining failed, command={0}, stderr={1}, returncode={2}'.format(cmddebug, stderr, proc.returncode))
+            logger.error('pdftk joining failed, command={0}, stderr={1}, returncode={2}'.format(cmddebug, stderr, proc.returncode))
 
     except Exception as ex:
-        log.error('pdftk joining failed, command={0}, exception={1}, stderr={2}'.format(cmddebug, ex, stderr))
+        logger.error('pdftk joining failed, command={0}, exception={1}, stderr={2}'.format(cmddebug, ex, stderr))
         raise
 
     return stdout
@@ -236,7 +274,7 @@ def pdf_set_metadata(pdf_payload, metadata):
 
     command = [find_pdftk(), '-', 'update_info', tmpfile.name, 'output', '-']
 
-    #log.info('command={0}'.format(' '.join(command)))
+    #logger.info('command={0}'.format(' '.join(command)))
 
     cmddebug = ' '.join(command)
     stdout = stderr = ''
@@ -254,11 +292,11 @@ def pdf_set_metadata(pdf_payload, metadata):
 
         stdout, stderr = proc.communicate(pdf_payload)
         if proc.returncode is not None and proc.returncode != 0:
-            log.error('pdftk metadata store failed, command={0}, stderr={1}'.format(cmddebug, stderr))
+            logger.error('pdftk metadata store failed, command={0}, stderr={1}'.format(cmddebug, stderr))
             raise Exception()
 
     except Exception as ex:
-        log.error('pdftk metadata store failed, command={0}, exception={1}, stderr={2}'.format(cmddebug, ex, stderr))
+        logger.error('pdftk metadata store failed, command={0}, exception={1}, stderr={2}'.format(cmddebug, ex, stderr))
         raise
 
     return stdout
@@ -326,20 +364,34 @@ def find_convert():
     Debian: aptitude install imagemagick
     /usr/bin/convert
 
-    Mac OS X
+    Mac OS X with Homebrew
+    /usr/local/bin/convert
+
+    Mac OS X with Macports
     /opt/local/bin/convert
 
     Self-compiled
+    /opt/imagemagick/bin/convert
     /opt/imagemagick-7.0.2/bin/convert
     """
 
+    # Some nailed location candidates
     candidates = [
         '/opt/imagemagick-7.0.2/bin/convert',
         '/opt/imagemagick/bin/convert',
+        '/usr/local/bin/convert',
         '/opt/local/bin/convert',
         '/usr/bin/convert',
-        ]
-    return find_program_candidate(candidates)
+    ]
+
+    # More location candidates from the system
+    candidates += where.where('convert')
+
+    # Find location of "convert" program
+    convert_path = find_program_candidate(candidates)
+
+    logger.info('Found "convert" program at {}'.format(convert_path))
+    return convert_path
 
 @memoize
 def find_pdftk():
@@ -359,6 +411,10 @@ def find_pdftk():
         '/usr/local/bin/pdftk',
         '/usr/bin/pdftk',
     ]
+
+    # More location candidates from the system
+    candidates += where.where('pdftk')
+
     return find_program_candidate(candidates)
 
 def find_program_candidate(candidates):