+++ CHANGES.rst
@@ -15,6 +15,7 @@ development
 - [mw] Direct access to OPS 3.1 API has been disabled, so stop linking to it
 - [ui] Fix: When creating new project, convert project name to string
 - [mw] Improve DpmaRegisterAccess re. language selection
+- [mw] Improve image conversion by using the Pillow library
 
 2018-03-17 0.150.0
 ==================